requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.2
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==2.0
django-tables2==2.5.3
pydantic==1.10.7
sqlparse==0.4.4
typing_extensions==4.5.0
